rename : entity enum(gender) 위치 변경 #20
entity -> entity.enums
