+++ src/main/java/com/openvelog/openvelogbe/common/entity/Member.java
@@ -1,5 +1,6 @@
 package com.openvelog.openvelogbe.common.entity;
 
+import com.openvelog.openvelogbe.common.entity.enums.Gender;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
